api/requirements2.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
aniso8601=8.0.0=py_0
backcall=0.1.0=py37_0
blas=1.0=mkl
boto3=1.12.39=pyh9f0ad1d_0
botocore=1.15.39=pyh9f0ad1d_0
ca-certificates=2020.1.1=0
certifi=2020.4.5.1=py37_0
cffi=1.14.0=py37h7a1dbc1_0
chardet=3.0.4=py37hc8dfbb8_1006
click=7.1.1=pyh8c360ce_0
colorama=0.4.3=py_0
console_shortcut=0.1.1=4
cpuonly=1.0=0
cryptography=2.8=py37hb32ad35_1
decorator=4.4.2=py_0
docutils=0.15.2=py37_0
entrypoints=0.3=py37_0
flask=1.1.2=pyh9f0ad1d_0
flask-restful=0.3.8=py_0
freetype=2.9.1=ha9979f8_1
icc_rt=2019.0.0=h0cc432a_1
icu=58.2=ha66f8fd_1
idna=2.9=py_1
intel-openmp=2020.0=166
ipykernel=5.1.4=py37h39e3cac_0
ipython=7.13.0=py37h5ca1d4c_0
ipython_genutils=0.2.0=py37_0
itsdangerous=1.1.0=py_0
jedi=0.16.0=py37_1
jinja2=2.11.1=py_0
jmespath=0.9.5=py_0
joblib=0.14.1=py_0
jpeg=9b=hb83a4c4_2
jupyter_client=6.1.2=py_0
jupyter_core=4.6.3=py37_0
libpng=1.6.37=h2a8f88b_0
libsodium=1.0.16=h9d3ae62_0
libtiff=4.1.0=h56a325e_0
markupsafe=1.1.1=py37h8055547_1
mkl=2020.0=166
mkl-service=2.3.0=py37hb782905_0
mkl_fft=1.0.15=py37h14836fe_0
mkl_random=1.1.0=py37h675688f_0
ninja=1.9.0=py37h74a9793_0
numpy=1.18.1=py37h93ca92e_0
numpy-base=1.18.1=py37hc3f5095_1
olefile=0.46=py37_0
openssl=1.1.1f=he774522_0
parso=0.6.2=py_0
pickleshare=0.7.5=py37_0
pillow=7.0.0=py37hcc1f983_0
pip=20.0.2=py37_1
prompt-toolkit=3.0.4=py_0
prompt_toolkit=3.0.4=0
pycparser=2.20=py_0
pygments=2.6.1=py_0
pyopenssl=19.1.0=py_1
pyqt=5.9.2=py37h6538335_2
pysocks=1.7.1=py37hc8dfbb8_1
python=3.7.7=h60c2a47_0_cpython
python-dateutil=2.8.1=py_0
python_abi=3.7=1_cp37m
pytorch=1.4.0=py3.7_cpu_0
pytz=2019.3=py_0
pywin32=227=py37he774522_1
pyzmq=18.1.1=py37ha925a31_0
qt=5.9.7=vc14h73c81de_0
qtconsole=4.7.2=py_0
qtpy=1.9.0=py_0
regex=2020.4.4=py37h8055547_0
requests=2.23.0=pyh8c360ce_2
s3transfer=0.3.3=py37hc8dfbb8_1
sacremoses=0.0.38=py_0
setuptools=46.1.3=py37_0
sip=4.19.8=py37h6538335_0
six=1.14.0=py37_0
sqlite=3.31.1=he774522_0
tk=8.6.8=hfa6e2cd_0
torchvision=0.5.0=py37_cpu
tornado=6.0.4=py37he774522_1
tqdm=4.45.0=pyh9f0ad1d_0
traitlets=4.3.3=py37_0
transformers=2.1.1=py_0
urllib3=1.25.8=py37hc8dfbb8_1
vc=14.1=h0510ff6_4
vs2015_runtime=14.16.27012=hf0eaf9b_1
wcwidth=0.1.9=py_0
werkzeug=1.0.1=pyh9f0ad1d_0
wheel=0.34.2=py37_0
win_inet_pton=1.1.0=py37_0
wincertstore=0.2=py37_0
xz=5.2.4=h2fa13f4_4
zeromq=4.3.1=h33f27b4_3
zlib=1.2.11=h62dcd97_3
zstd=1.3.7=h508b16e_0
